+++ SBO-Lib/lib/SBO/Lib/Util.pm
@@ -17,7 +17,7 @@ use 5.016;
 use strict;
 use warnings;
 
-our $VERSION = '4.0.1';
+our $VERSION = '4.1';
 
 use Exporter 'import';
 use File::Copy;