+++ SBO-Lib/ChangeLog.md
@@ -1,6 +1,6 @@
 # sbotools
 
-#### Post-Release
+#### 4.1 - 2025-10-31
   * Drop support for Slackware 14.0, 14.1 and 14.2
   * sbocheck, sbotool: Add a perl package compatibility test
   * sbocheck: --obsolete-check also downloads the perl version history file